refactor: tests now using txProposal new format
