+++ tests/txProposal.test.ts
@@ -3,7 +3,6 @@ import { send as txProposalSend } from '@src/api/txProposalSend';
 import { destroy as txProposalDestroy } from '@src/api/txProposalDestroy';
 import {
   getTxProposal,
-  getTxProposalOutputs,
   getUtxos,
   updateTxProposal,
   getTxProposalTokenInfo,
@@ -21,7 +20,6 @@ import {
   ADDRESSES,
 } from '@tests/utils';
 import { APIGatewayProxyResult } from 'aws-lambda';
-import buffer from 'buffer';
 
 import { ApiError } from '@src/api/errors';
 
@@ -40,13 +38,12 @@ afterAll(async () => {
   await closeDbConnection(mysql);
 });
 
-const _checkTxProposalTables = async (txProposalId, inputs, outputs): Promise<void> => {
+const _checkTxProposalTables = async (txProposalId, inputs): Promise<void> => {
   const utxos = await getUtxos(mysql, inputs);
   for (const utxo of utxos) {
     expect(utxo.txProposalId).toBe(txProposalId);
   }
   expect(await getTxProposal(mysql, txProposalId)).not.toBeNull();
-  expect(await getTxProposalOutputs(mysql, txProposalId)).toStrictEqual(outputs);
 };
 
 const _checkTxProposalInfoTable = async (txProposalId, name, symbol): Promise<void> => {
@@ -284,14 +281,7 @@ test('PUT /txproposals/{proposalId} with a missing proposalId should fail with A
   expect.hasAssertions();
 
   const txSendEvent = makeGatewayEventWithAuthorizer('my-wallet', { txProposalId: '8d1e2921-7bc9-41f5-9758-40b734edff0f' }, JSON.stringify({
-    inputsSignatures: [1],
-    nonce: 28,
-    parents: [
-      '00000000204080e1b7563558869e39d169efae5008d2158cc3cb6c0c3812ff2a',
-      '00000000414ede4aad9e08e3a336191e2b0510d9f74c7b5e94b68e653bcbf42e',
-    ],
-    timestamp: 1609881763,
-    weight: 19,
+    txHex: 'txhex',
   }));
   const txSendResult = await txProposalSend(txSendEvent, null, null) as APIGatewayProxyResult;
 
@@ -378,124 +368,11 @@ test('PUT /txproposals/{proposalId} on a proposal which status is not OPEN or SE
   );
 
   const txSendEvent = makeGatewayEventWithAuthorizer('my-wallet', { txProposalId: returnBody.txProposalId }, JSON.stringify({
-    inputsSignatures: [1],
-    nonce: 28,
-    parents: [
-      '00000000204080e1b7563558869e39d169efae5008d2158cc3cb6c0c3812ff2a',
-      '00000000414ede4aad9e08e3a336191e2b0510d9f74c7b5e94b68e653bcbf42e',
-    ],
-    timestamp: 1609881763,
-    weight: 19,
+    txHex,
   }));
   const txSendResult = await txProposalSend(txSendEvent, null, null) as APIGatewayProxyResult;
 
   expect(JSON.parse(txSendResult.body as string).error).toStrictEqual(ApiError.TX_PROPOSAL_NOT_OPEN);
-});
-
-test('PUT /txproposals/{proposalId} with an invalid weight should fail with ApiError.INVALID_TX_WEIGHT', async () => {
-  expect.hasAssertions();
-
-  // Create the spy to mock wallet-lib
-  const spy = jest.spyOn(hathorLib.axios, 'createRequestInstance');
-  spy.mockReturnValue({
-    post: () => Promise.resolve({
-      data: { success: true },
-    }),
-    get: () => Promise.resolve({
-      data: {
-        success: true,
-        version: '0.38.0',
-        network: 'mainnet',
-        min_weight: 14,
-        min_tx_weight: 14,
-        min_tx_weight_coefficient: 1.6,
-        min_tx_weight_k: 100,
-        token_deposit_percentage: 0.01,
-        reward_spend_min_blocks: 300,
-        max_number_inputs: 255,
-        max_number_outputs: 255,
-      },
-    }),
-  });
-
-  await addToWalletTable(mysql, [['my-wallet', 'xpubkey', 'ready', 5, 10000, 10001]]);
-  await addToAddressTable(mysql, [{
-    address: ADDRESSES[0],
-    index: 0,
-    walletId: 'my-wallet',
-    transactions: 2,
-  }]);
-
-  const token1 = '004d75c1edd4294379e7e5b7ab6c118c53c8b07a506728feb5688c8d26a97e50';
-  const token2 = '002f2bcc3261b4fb8510a458ed9df9f6ba2a413ee35901b3c5f81b0c085287e2';
-
-  const utxos = [
-    ['004d75c1edd4294379e7e5b7ab6c118c53c8b07a506728feb5688c8d26a97e50', 0, token1, ADDRESSES[0], 300, 0, null, null, false],
-    ['0000001e39bc37fe8710c01cc1e8c0a937bf6f9337551fbbfddc222bfc28c197', 0, token1, ADDRESSES[0], 100, 0, null, null, false],
-    ['00000060a25077e48926bcd9473d77259296e123ec6af1c1a16c1c381093ab90', 0, token2, ADDRESSES[0], 300, 0, null, null, false],
-  ];
-
-  await addToUtxoTable(mysql, utxos);
-  await addToWalletBalanceTable(mysql, [{
-    walletId: 'my-wallet',
-    tokenId: 'token1',
-    unlockedBalance: 400,
-    lockedBalance: 0,
-    unlockedAuthorities: 0,
-    lockedAuthorities: 0,
-    timelockExpires: null,
-    transactions: 2,
-  }, {
-    walletId: 'my-wallet',
-    tokenId: 'token2',
-    unlockedBalance: 300,
-    lockedBalance: 0,
-    unlockedAuthorities: 0,
-    lockedAuthorities: 0,
-    timelockExpires: null,
-    transactions: 1,
-  }]);
-
-  await addToAddressTable(mysql, [{
-    address: ADDRESSES[1],
-    index: 1,
-    walletId: 'my-wallet',
-    transactions: 0,
-  }]);
-
-  // only one output, spending the whole 300 utxo of token1
-  const outputs = [new hathorLib.Output(300, new hathorLib.Address(ADDRESSES[0], { network: new hathorLib.Network(process.env.NETWORK) }), {
-    tokenData: 1,
-  })];
-  const inputs = [new hathorLib.Input(utxos[0][0], utxos[0][1])];
-  const transaction = new hathorLib.Transaction(inputs, outputs, { tokens: [token1] });
-
-  const txHex = transaction.toHex();
-  const event = makeGatewayEventWithAuthorizer('my-wallet', null, JSON.stringify({ txHex }));
-  const txCreateResult = await txProposalCreate(event, null, null) as APIGatewayProxyResult;
-  const returnBody = JSON.parse(txCreateResult.body as string);
-
-  const signature = buffer.Buffer.alloc(20);
-  const pubkeyBytes = buffer.Buffer.alloc(30);
-
-  const txSendEvent = makeGatewayEventWithAuthorizer('my-wallet', { txProposalId: returnBody.txProposalId }, JSON.stringify({
-    inputsSignatures: [
-      1, 2, 3, 4, 5, 6, 7,
-    ].map(() => hathorLib.transaction.createInputData(signature, pubkeyBytes).toString('base64')),
-    nonce: 28,
-    parents: [
-      '00000000204080e1b7563558869e39d169efae5008d2158cc3cb6c0c3812ff2a',
-      '00000000414ede4aad9e08e3a336191e2b0510d9f74c7b5e94b68e653bcbf42e',
-    ],
-    timestamp: 1609881763,
-    weight: -1,
-  }));
-  const txSendResult = await txProposalSend(txSendEvent, null, null) as APIGatewayProxyResult;
-
-  expect(JSON.parse(txSendResult.body).success).toStrictEqual(false);
-  expect(JSON.parse(txSendResult.body).error).toStrictEqual(ApiError.INVALID_TX_WEIGHT);
-
-  spy.mockRestore();
 });
 
 test('PUT /txproposals/{proposalId} with an invalid txHex should fail and update tx_proposal to SEND_ERROR', async () => {
@@ -584,20 +461,8 @@ test('PUT /txproposals/{proposalId} with an invalid txHex should fail and update
   const txCreateResult = await txProposalCreate(event, null, null) as APIGatewayProxyResult;
   const returnBody = JSON.parse(txCreateResult.body as string);
 
-  const signature = buffer.Buffer.alloc(20);
-  const pubkeyBytes = buffer.Buffer.alloc(30);
-
   const txSendEvent = makeGatewayEventWithAuthorizer('my-wallet', { txProposalId: returnBody.txProposalId }, JSON.stringify({
-    inputsSignatures: [
-      1, 2, 3, 4, 5, 6, 7,
-    ].map(() => hathorLib.transaction.createInputData(signature, pubkeyBytes).toString('base64')),
-    nonce: 28,
-    parents: [
-      '00000000204080e1b7563558869e39d169efae5008d2158cc3cb6c0c3812ff2a',
-      '00000000414ede4aad9e08e3a336191e2b0510d9f74c7b5e94b68e653bcbf42e',
-    ],
-    timestamp: 1609881763,
-    weight: 16.74,
+    txHex,
   }));
   const txSendResult = await txProposalSend(txSendEvent, null, null) as APIGatewayProxyResult;
 
@@ -693,20 +558,8 @@ test('PUT /txproposals/{proposalId} should update tx_proposal to SEND_ERROR on f
   const txCreateResult = await txProposalCreate(event, null, null) as APIGatewayProxyResult;
   const returnBody = JSON.parse(txCreateResult.body as string);
 
-  const signature = buffer.Buffer.alloc(20);
-  const pubkeyBytes = buffer.Buffer.alloc(30);
-
   const txSendEvent = makeGatewayEventWithAuthorizer('my-wallet', { txProposalId: returnBody.txProposalId }, JSON.stringify({
-    inputsSignatures: [
-      1, 2, 3, 4, 5, 6, 7,
-    ].map(() => hathorLib.transaction.createInputData(signature, pubkeyBytes).toString('base64')),
-    nonce: 28,
-    parents: [
-      '00000000204080e1b7563558869e39d169efae5008d2158cc3cb6c0c3812ff2a',
-      '00000000414ede4aad9e08e3a336191e2b0510d9f74c7b5e94b68e653bcbf42e',
-    ],
-    timestamp: 1609881763,
-    weight: 16.74,
+    txHex,
   }));
   const txSendResult = await txProposalSend(txSendEvent, null, null) as APIGatewayProxyResult;
 
@@ -912,7 +765,7 @@ test('POST /txproposals one output and input on txHex', async () => {
   expect(returnBody.outputs).toContainEqual({ address: ADDRESSES[0], value: 300, token: token1, tokenData: 1, timelock: null });
   expect(returnBody.tokens).toContainEqual(token1);
 
-  await _checkTxProposalTables(returnBody.txProposalId, returnBody.inputs, returnBody.outputs);
+  await _checkTxProposalTables(returnBody.txProposalId, returnBody.inputs);
 });
 
 test('POST /txproposals with denied utxos', async () => {
@@ -1031,7 +884,6 @@ test('POST /txproposals a tx create action on txHex', async () => {
   expect(returnBody.outputs).toHaveLength(4);
   expect(returnBody.outputs).toContainEqual({ address: ADDRESSES[0], value: 200, token: '00', tokenData: 0, timelock: null });
 
-  // await _checkTxProposalTables(returnBody.txProposalId, returnBody.inputs, returnBody.outputs);
   await _checkTxProposalInfoTable(returnBody.txProposalId, name, symbol);
 });
 
@@ -1122,20 +974,9 @@ test('PUT /txproposals/{proposalId} with txhex', async () => {
   const result = await txProposalCreate(event, null, null) as APIGatewayProxyResult;
   const returnBody = JSON.parse(result.body as string);
 
-  const signature = buffer.Buffer.alloc(20);
-  const pubkeyBytes = buffer.Buffer.alloc(30);
-
   const txSendEvent = makeGatewayEventWithAuthorizer('my-wallet', { txProposalId: returnBody.txProposalId }, JSON.stringify({
-    inputsSignatures: [1].map(() => hathorLib.transaction.createInputData(signature, pubkeyBytes).toString('base64')),
-    nonce: 28,
-    parents: [
-      '00000000204080e1b7563558869e39d169efae5008d2158cc3cb6c0c3812ff2a',
-      '00000000414ede4aad9e08e3a336191e2b0510d9f74c7b5e94b68e653bcbf42e',
-    ],
-    timestamp: 1609881763,
-    weight: 16.74,
+    txHex,
   }));
-
   const txSendResult = await txProposalSend(txSendEvent, null, null) as APIGatewayProxyResult;
 
   const sendReturnBody = JSON.parse(txSendResult.body as string);
